+++ R/coerce.R
@@ -75,6 +75,7 @@ setMethod(
     if (length(groups) == 1) {
       if (is.character(groups)) groups <- match(groups, cols)
       grp <- as.character(from[[groups]])
+      grp[grp == ""] <- NA_character_
     }
     n_grp <- length(unique(grp[!is.na(grp)]))
 
@@ -197,14 +198,17 @@ as.data.frame.LogRatio <- function(x, ...) {
 #' @method as.data.frame OutlierIndex
 #' @export
 as.data.frame.OutlierIndex <- function(x, ...) {
-  y <- methods::S3Part(x, strictS3 = TRUE)
+  out <- methods::as(x, "matrix")
+  colnames(out) <- paste0("out_", colnames(out))
+  d <- x@distances
+  colnames(d) <- paste0("dist_", colnames(d))
   data.frame(
-    index = seq_along(y),
+    index = seq_len(nrow(out)),
     sample = get_samples(x),
     group = get_groups(x),
-    distance = x@distances,
-    outlier = y,
-    row.names = NULL,
+    d,
+    out,
+    row.names = get_identifiers(x),
     stringsAsFactors = FALSE
   )
 }