+++ R/plot.R
@@ -3,6 +3,7 @@
 NULL
 
 # CompositionMatrix ============================================================
+## Plot ------------------------------------------------------------------------
 #' @export
 #' @method plot CompositionMatrix
 plot.CompositionMatrix <- function(x, ..., margin = NULL) {
@@ -15,6 +16,108 @@ plot.CompositionMatrix <- function(x, ..., margin = NULL) {
 #' @aliases plot,CompositionMatrix,missing-method
 setMethod("plot", c(x = "CompositionMatrix", y = "missing"), plot.CompositionMatrix)
 
+## Histogram -------------------------------------------------------------------
+#' @export
+#' @method hist CompositionMatrix
+hist.CompositionMatrix <- function(x, ..., freq = FALSE,
+                                   ncol = NULL, flip = FALSE,
+                                   main = NULL, sub = NULL,
+                                   ann = graphics::par("ann"),
+                                   axes = TRUE, frame.plot = axes) {
+  m <- nrow(x)
+  p <- ncol(x)
+
+  # TODO
+  ilr <- TRUE
+
+  ## Plot
+  if (is.null(ncol)) ncol <- if (p > 4) 2 else 1
+  nrow <- ceiling(p / ncol)
+
+  ## Graphical parameters
+  ## Save and restore
+  if (p > 1) {
+    old_par <- graphics::par(
+      mar = c(4.1, 5.1, 4.1, if (flip) 5.1 else 2.1),
+      oma = c(0, 0, 5, 0),
+      mfcol = c(nrow, ncol)
+    )
+    on.exit(graphics::par(old_par))
+  }
+
+  cex.axis <- list(...)$cex.axis %||% graphics::par("cex.axis")
+  col.axis <- list(...)$col.axis %||% graphics::par("col.axis")
+  font.axis <- list(...)$font.axis %||% graphics::par("font.axis")
+  cex.lab <- list(...)$cex.lab %||% graphics::par("cex.lab")
+  if (p > 1) cex.lab <- cex.lab * ifelse(max(m, p) < 3, 0.83,  0.66) # See ?par
+  col.lab <- list(...)$col.lab %||% graphics::par("col.lab")
+  font.lab <- list(...)$font.lab %||% graphics::par("font.lab")
+  cex.main <- list(...)$cex.main %||% graphics::par("cex.main")
+  font.main <- list(...)$font.main %||% graphics::par("font.main")
+  col.main <- list(...)$col.main %||% graphics::par("col.main")
+
+  index <- seq_len(p)
+  for (j in index) {
+    ## Compute univariate ilr transformation
+    xi <- x[, j]
+    zi <- sqrt(1 / 2) * log(xi / (1 - xi))
+
+    lab_i <- pretty(xi, n = 6)
+    lab_i <- lab_i[lab_i > 0]
+    at_i <- sqrt(1 / 2) * log(lab_i / (1 - lab_i))
+
+    ## Histogram
+    h <- graphics::hist(x = zi, ..., plot = FALSE)
+    xlim <- range(at_i, h$breaks, finite = TRUE)
+    plot(h, freq = freq, xlim = xlim,
+         main = NULL, sub = NULL, xlab = NULL, ylab = NULL, axes = FALSE)
+
+    ## Construct axis
+    y_side <- if (j %% 2 || !flip) 2 else 4
+    if (axes) {
+      graphics::axis(side = 1, cex.axis = cex.axis, col.axis = col.axis,
+                     font.axis = font.axis, xpd = NA, las = 1)
+      graphics::axis(side = 3, at = at_i, labels = lab_i,
+                     cex.axis = cex.axis, col.axis = col.axis,
+                     font.axis = font.axis, xpd = NA, las = 1)
+      graphics::axis(side = y_side, cex.axis = cex.axis, col.axis = col.axis,
+                     font.axis = font.axis, xpd = NA, las = 1)
+    }
+
+    ## Plot frame
+    if (frame.plot) {
+      graphics::box()
+    }
+
+    ## Add annotation
+    if (ann) {
+      xlab <- colnames(x)[j]
+      ylab <- "Frequency"
+      graphics::mtext(sprintf("ilr(%s)", xlab), side = 1, line = 2.5,
+                      cex = cex.lab, col = col.lab, font = font.lab)
+      graphics::mtext(sprintf("%s %%", xlab), side = 3, line = 2.5,
+                      cex = cex.lab, col = col.lab, font = font.lab)
+      graphics::mtext(ylab, side = y_side, line = 3, cex = cex.lab,
+                      col = col.lab, font = font.lab)
+    }
+  }
+
+  ## Add annotation
+  if (ann) {
+    graphics::par(mfcol = c(1, 1))
+    graphics::mtext(main, side = 3, line = 3, cex = cex.main, font = font.main,
+                    col = col.main)
+  }
+
+  invisible(x)
+}
+
+#' @export
+#' @rdname hist
+#' @aliases hist,CompositionMatrix-method
+setMethod("hist", c(x = "CompositionMatrix"), hist.CompositionMatrix)
+
+## Barplot ---------------------------------------------------------------------
 #' @export
 #' @method barplot CompositionMatrix
 barplot.CompositionMatrix <- function(height, ...,
@@ -22,18 +125,18 @@ barplot.CompositionMatrix <- function(height, ...,
                                       groups = get_groups(height), horiz = TRUE,
                                       xlab = NULL, ylab = NULL,
                                       main = NULL, sub = NULL,
-                                      ann = graphics::par("ann"), axes = TRUE) {
+                                      ann = graphics::par("ann"), axes = TRUE,
+                                      col = grDevices::hcl.colors(ncol(height), "viridis"),
+                                      legend = list()) {
   ## Get data
   z <- height@.Data
 
   ## Ordering
-  if (!is.null(order)) {
-    ordering <- order(z[, order], decreasing = decreasing)
-    z <- z[ordering, ]
-  }
+  ordering <- seq_len(nrow(height))
+  if (!is.null(order)) ordering <- order(z[, order], decreasing = decreasing)
+  z <- z[ordering, ]
 
   ## Graphical parameters
-  col <- list(...)$col %||% grDevices::hcl.colors(ncol(z), "viridis")
   cex.axis <- list(...)$cex.axis %||% graphics::par("cex.axis")
   col.axis <- list(...)$col.axis %||% graphics::par("col.axis")
   font.axis <- list(...)$font.axis %||% graphics::par("font.axis")
@@ -45,9 +148,11 @@ barplot.CompositionMatrix <- function(height, ...,
   col.main <- list(...)$col.main %||% graphics::par("col.main")
 
   ## Grouping
+  n <- 0
   if (length(stats::na.omit(groups)) > 0) {
     arkhe::assert_length(groups, nrow(z))
 
+    groups <- groups[ordering]
     groups <- factor(groups, exclude = NULL)
     grp <- split(as.data.frame(z), f = groups)
     n <- nlevels(groups)
@@ -92,13 +197,30 @@ barplot.CompositionMatrix <- function(height, ...,
     ## Add annotation
     if (ann) {
       graphics::par(mfcol = c(1, 1))
-      graphics::mtext(main, side = 3, line = 3, cex = cex.main, font = font.main,
-                      col = col.main)
+      graphics::mtext(main, side = 3, line = 3, cex = cex.main,
+                      font = font.main, col = col.main)
     }
   } else {
     graphics::barplot(height = t(z), horiz = horiz, col = col, las = 1,
                       main = main, sub = sub, xlab = xlab, ylab = ylab,
                       axes = axes, ann = ann, ...)
+  }
+
+  ## Add legend
+  # https://stackoverflow.com/a/42076830
+  if (is.list(legend)) {
+    leg_par <- graphics::par(mar = c(0, 0, 0, 0), oma = c(0, 0, 0, 0),
+                             mfcol = c(1, 1), new = TRUE)
+    on.exit(graphics::par(leg_par), add = TRUE)
+    graphics::plot(0, 0, type = "n", ann = FALSE, axes = FALSE)
+
+    ## Compute legend position
+    args <- list(x = "top", legend = colnames(height), fill = col,
+                 ncol = ceiling(ncol(height) / 2), bty = "n", xpd = NA)
+    args <- utils::modifyList(args, legend)
+
+    ## Plot legend
+    do.call(graphics::legend, args = args)
   }
 
   invisible(height)