+++ R/statistics.R
@@ -2,28 +2,54 @@
 #' @include AllGenerics.R
 NULL
 
+# Split ========================================================================
+#' @export
+#' @method split CompositionMatrix
+split.CompositionMatrix <- function(x, f, drop = FALSE, ...) {
+  lapply(
+    X = split(x = seq_len(nrow(x)), f = f, drop = drop, sep = "_", ...),
+    FUN = function(ind) x[ind, , drop = FALSE]
+  )
+}
+
+#' @export
+#' @rdname split
+#' @aliases split,CompositionMatrix-method
+setMethod("split", "CompositionMatrix", split.CompositionMatrix)
+
+#' @export
+#' @method split LogRatio
+split.LogRatio <- function(x, f, drop = FALSE, ...) {
+  lapply(
+    X = split(x = seq_len(nrow(x)), f = f, drop = drop, sep = "_", ...),
+    FUN = function(ind) x[ind, , drop = FALSE]
+  )
+}
+
+#' @export
+#' @rdname split
+#' @aliases split,LogRatio-method
+setMethod("split", "LogRatio", split.LogRatio)
+
 # Aggregate ====================================================================
 #' @export
 #' @method aggregate CompositionMatrix
-aggregate.CompositionMatrix <- function(x, by, FUN, ...) {
-  ## Validation
-  by <- match.arg(by, choices = c("samples", "groups"), several.ok = FALSE)
+aggregate.CompositionMatrix <- function(x, by, FUN, ...,
+                                        simplify = TRUE, drop = TRUE) {
+  m <- nrow(x)
 
-  if (by == "samples") {
-    if (!any_replicated(x)) {
-      warning("No observations are repeated.", call. = FALSE)
-    }
-    index <- get_samples(x)
-  }
-  if (by == "groups") {
-    if (!any_assigned(x)) {
-      stop("No group is defined.", call. = FALSE)
-    }
-    index <- get_groups(x)
+  ## Validation
+  if (!is.list(by)) by <- list(by)
+  arkhe::assert_lengths(by, m)
+
+  index <- interaction(by, drop = drop, sep = "_")
+  if (length(unique(by)) == m) {
+    warning("Nothing to group by.", call. = FALSE)
+    return(x)
   }
 
   m <- tapply(
-    X = seq_len(nrow(x)),
+    X = seq_len(m),
     INDEX = index,
     FUN = function(i, data, fun, ...) fun(data[i, , drop = FALSE], ...),
     data = x,
@@ -31,6 +57,14 @@ aggregate.CompositionMatrix <- function(x, by, FUN, ...) {
     ...,
     simplify = FALSE
   )
+
+  has_dim <- vapply(
+    X = m,
+    FUN = function(x) !is.null(nrow(x)) && nrow(x) > 1,
+    FUN.VALUE = logical(1)
+  )
+
+  if (any(has_dim) || !simplify) return(m)
   do.call(rbind, m)
 }
 
@@ -93,9 +127,31 @@ setMethod("mean", "CompositionMatrix", mean.CompositionMatrix)
 #' @keywords internal
 #' @noRd
 gmean <- function(x, trim = 0, na.rm = FALSE) {
-  index <- is.finite(x) & x > 0
-  exp(mean(log(unclass(x)[index]), trim = trim, na.rm = na.rm))
+  if (na.rm) x <- x[is.finite(x)]
+  x <- x[x > 0]
+  exp(mean(log(x), trim = trim))
 }
+
+# Quantile =====================================================================
+#' @export
+#' @method quantile CompositionMatrix
+quantile.CompositionMatrix <- function(x, ..., probs = seq(0, 1, 0.25),
+                                       na.rm = FALSE, names = TRUE) {
+  apply(
+    X = x,
+    MARGIN = 2,
+    FUN = stats::quantile,
+    probs = probs,
+    na.rm = na.rm,
+    names = names,
+    ...
+  )
+}
+
+#' @export
+#' @rdname quantile
+#' @aliases quantile,CompositionMatrix-method
+setMethod("quantile", "CompositionMatrix", quantile.CompositionMatrix)
 
 # Scale ========================================================================
 #' @export