+++ inst/tinytest/test_plot.R
@@ -13,6 +13,10 @@ if (at_home()) {
   plot_pairs <- function() plot(coda, order = NULL)
   expect_snapshot_plot(plot_pairs, "plot_pairs")
 
+  # Histogram ==================================================================
+  plot_hist <- function() hist(coda, ncol = 3)
+  expect_snapshot_plot(plot_hist, "plot_hist")
+
   # Barplot ====================================================================
   plot_barplot <- function() barplot(coda, order = NULL)
   expect_snapshot_plot(plot_barplot, "plot_barplot")
@@ -21,7 +25,7 @@ if (at_home()) {
   expect_snapshot_plot(plot_barplot_order, "plot_barplot_order")
 
   set_groups(coda) <- rep(1:5, 5)
-  plot_barplot_group <- function() barplot(coda, order = NULL)
+  plot_barplot_group <- function() barplot(coda, order = 2)
   expect_snapshot_plot(plot_barplot_group, "plot_barplot_group")
 
   plot_barplot_vertical <- function() barplot(coda, order = NULL, horiz = FALSE)